# packages/admin-api-client/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ReviewBatchResponse": {
        "additionalProperties": false,
        "properties": {
          "activeLearningVersion": {
            "title": "Activelearningversion",
            "type": "string"
          },
          "calibrationReportSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Calibrationreportsha256",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "datasetSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Datasetsha256",
            "type": "string"
          },
          "gameId": {
            "format": "uuid",
            "title": "Gameid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inventorySha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Inventorysha256",
            "type": "string"
          },
          "itemCount": {
            "title": "Itemcount",
            "type": "integer"
          },
          "modelArtifactSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Modelartifactsha256",
            "type": "string"
          },
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "sourceReportSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sourcereportsha256",
            "type": "string"
          },
          "splitSha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Splitsha256",
            "type": "string"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "id",
          "gameId",
          "sourceReportSha256",
          "activeLearningVersion",
          "modelVersion",
          "modelArtifactSha256",
          "calibrationReportSha256",
          "datasetSha256",
          "splitSha256",
          "inventorySha256",
          "temperature",
          "itemCount",
          "createdAt"
        ],
        "title": "ReviewBatchResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Game Predictor Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/review-batches": {
      "get": {
        "operationId": "listReviewBatches",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewBatchResponse"
                  },
                  "title": "Response Listreviewbatches",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List immutable manual-review batches",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "importReviewBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewBatchImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewBatchImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Idempotently import one active-learning selection",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-batches/{review_batch_id}": {
      "get": {
        "operationId": "getReviewBatch",
        "parameters": [
          {
            "in": "path",
            "name": "review_batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get immutable manual-review batch metadata",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-batches/{review_batch_id}/feedback-exports": {
      "get": {
        "operationId": "listReviewFeedbackExports",
        "parameters": [
          {
            "in": "path",
            "name": "review_batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewFeedbackExportResponse"
                  },
                  "title": "Response Listreviewfeedbackexports",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List immutable labeled-feedback exports for one review batch",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "createReviewFeedbackExport",
        "parameters": [
          {
            "in": "path",
            "name": "review_batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewFeedbackExportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFeedbackExportCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Create an immutable versioned labeled-feedback export",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-batches/{review_batch_id}/items": {
      "get": {
        "operationId": "listReviewItems",
        "parameters": [
          {
            "in": "path",
            "name": "review_batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewItemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "after_selection_rank",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Selection Rank",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List one deterministic page of whole-layout review items",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-feedback-exports/{feedback_export_id}": {
      "get": {
        "operationId": "getReviewFeedbackExport",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFeedbackExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get one immutable labeled-feedback export",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-items/{review_item_id}": {
      "get": {
        "operationId": "getReviewItem",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get one immutable whole-layout review item",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-items/{review_item_id}/assets/board": {
      "get": {
        "operationId": "getReviewBoardAsset",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Read the canonical whole-board image for one review item",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-items/{review_item_id}/assets/cells/{cell_index}": {
      "get": {
        "operationId": "getReviewCellAsset",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Item Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "cell_index",
            "required": true,
            "schema": {
              "title": "Cell Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Read one canonical cell crop for a review item",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-items/{review_item_id}/assets/source": {
      "get": {
        "operationId": "getReviewSourceAsset",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Read the checksum-bound source image for one review item",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/admin/review-items/{review_item_id}/resolution": {
      "post": {
        "operationId": "resolveReviewItem",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewResolutionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResolutionCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Review resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Immutable review conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Atomically append one idempotent review resolution",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local Admin API",
      "url": "http://127.0.0.1:8000"
    }
  ]
}